Moved method into correct steps file
